[MER-936] Create new Inventories context for publishers
